Macro kat::run

macro_rules! run {
    (
        $global_data: ident, 
        $test_data: ident,
        $($data: tt)*
    ) => { ... };
}

Runs the tests.

This macro will create the module “kat_tests” and the test function “tests”. The macro will run any passed statements for every Test it could find. Inside the macro you get access to your global variables, as well as test specific variables. Depending on your IDE, you will see runnable highlighting above your macro invocation. The macro depends on the invocation of kat_cfg, global and test.

Invocation

 
// Set the path to the test file
kat_cfg!(path/to/my/file);
 
// Define yout global variables
global!{
    pub my_global: usize
}
 
// Define your test specific variables
test!{
    pub my_test_var: usize
}
 
// Now run your tests
run! {
    // First give a name to your variable, containing your globals
    global_var,
     
    // Then give a name to your current testcase variable, containing
    // specific test data
    test_case,
 
    // Now pass the statements you want to run
    println!("{}", global_var.my_global); // We can access the global variable, its behind a reference tho
    println!("{}", test_case.my_test_var); // In similar fashion, the test case. Here you get owned access
    my_super_expensive_function(); // Any function can be called.
    mymod::my_function(); // Also from other modules
    let x = 25; // Any statement can be run!
}
 

Panics

This macro will panic if:

The file wasn’t found, could not be opened or read

Parsing from TOML was unsuccessful