Crate kat

Testing framework for known answer tests.

This crate aims to drastically reduce the boilerplate code associated with rust tests, as well as to make tests easier to write and extend.

This framework splits its implementation into the test implementation and data, which is stored in .toml files.

Example

Toml file layout

The toml file must contain two sections, the global section and the test section or sections.

[global] // In this section global variables are defined
my_global_var = "This is a global variable"
 
[[test]] // In these sections we define test cases
id = 0
data = "This is data for test 0"
 
[[test]] // Multiple tests can be defined like this
id = 1
data = "This is data for test 1"

Writing the tests

Writing the tests is just as straight forward as writing the data

// In your test files global namespace (e.g tests/my_test.rs)
 
// This macro configures the filepath of your test file.
// The path will be interpreted, relative to the workspace root.
//
// In this case kat will look for the file:
// "WORKSPACE_ROOT/tests/data/my_data.toml"
//
// The file extension can be ommited, since we only support toml
kat_cfg!(tests/data/my_data);
 
// Now we define our global variables
global! {
 
// The name of the variable must be the same
// as the one defined in your data file.
// Kat internally uses serde to parse the key in
// your toml file as this variable.
 my_global_var: String
}
 
// We want to parse the data key
// in the test file, into this custom struct
struct MyData(String);
 
// Now we define our test specific variables
// The same conventions, as in the global! 
// macro do apply.
test! {
  id: usize,
   
  // To be able to parse the string,
  // we need to add a serde deserialize
  // attribute. See serde for more info on that.
  // Kat offers helper functions which can be used
  // in serdes deserialize_with attribute.
  #[serde(deserialize_with = "kat::from_string")]
  data: MyData
}
 
// Now we also need to implement From<String> for MyData,
// since the trait bounds on kat::from_string require us
// to do so
impl From<String> for MyData {
  fn from(s: String) -> Self {
    MyData(s)
  }
}
 
// And finally we provide the runner for our tests
run! {
    // The first two args of this macro are
    // the variable name for the global variables holder,
    // and the testcase variables holder.
    globals, test_case,
 
    // Now pass the statements you want to run
    println!("{}", global.my_global_var); // We can access the global variable, its behind a reference though
    println!("{}", test_case.id); // In similar fashion, the test case. Here you get owned access
    my_super_expensive_function(); // Any function can be called.
    mymod::my_function(); // Also from other modules
    let x = 25; // Any statement can be run!
 }
 

Modules

de

Helper functions to use in serde deserialize_with field attribute

Macros

global

Defines the global variables inside the test file.

kat_cfg

Configure the test files location.

run

Runs the tests.

test

Defines the test specific variables inside the test file.